add buildspec for wes_adapter

+++ wes_adapter/install.sh
@@ -1,9 +1,9 @@
 #! /bin/bash
 
-# if ! hash python3.9; then
-#     echo "python3.9 is not installed"
-#     exit 1
-# fi
+if ! hash python3.9; then
+    echo "python3.9 is not installed"
+    exit 1
+fi
 
 rm -rf ./dist && mkdir ./dist
 python -m pip install -r requirements.txt --target ./dist && (cd ./dist && zip -r ./wes_adapter.zip .)
